Move PostScheduler to separate file
